refactor Test-timeSplitting.C 1.3
